aurora/requirements.txt
pylint
pyyaml
clize
docker